test(block-editing): share no-reflow e2e helpers across both hosts
Add @quarto/preview-e2e-helpers (ts-packages/) holding the host-agnostic
inline-edit no-reflow assertions, parameterized by a Playwright FrameLocator
so the same logic runs against both hosts that embed @quarto/preview-renderer:

- q2-preview-spa (real q2 preview binary) -> uses assertNoReflowOnActivation
  (lean: active-region sizing via #q2-active-edit-region + following-sibling
  top, scroll-invariant).
- hub-client (Automerge + Monaco) -> uses measureLayout/expectLayoutStable
  (broad: all-blocks + document height) + activateBlock.

Both specs drop their local copies and import from the shared package; the
hub-client zero-reflow tests are behavior-preserving relocations (helpers moved
verbatim), keeping their decoration/left-column checks inline.

Lockfile: regenerated with npm 10 (the version matching the committed
lockfile format) to avoid npm 11.5.1 stripping esbuild optional platform
binaries; the diff is additive only (new workspace + devDeps), esbuild intact.
