+++ hub-client/package.json
@@ -55,6 +55,7 @@
     "@automerge/automerge-repo-storage-nodefs": "^2.5.6",
     "@eslint/js": "^9.39.1",
     "@playwright/test": "^1.60.0",
+    "@quarto/preview-e2e-helpers": "*",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@types/compression": "^1.8.1",
