+++ q2-preview-spa/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.50.0",
+    "@quarto/preview-e2e-helpers": "*",
     "@quarto/quarto-automerge-schema": "*",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
