+++ ts-packages/preview-e2e-helpers/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@quarto/preview-e2e-helpers",
+  "private": true,
+  "version": "0.0.1",
+  "type": "module",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "source": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@playwright/test": "^1.60.0",
+    "typescript": "~5.9.3"
+  }
+}
